Add permission settings to ANF SMB docs
A common user query is how to change the permission for a newly created SMB volume in Azure NetApp Files. The standard methods to change these permissions for a SMB share can be used. The idea is to add a short description in the "Create an SMB volume for Azure NetApp Files" page.